requirements-dev.txt
-r requirements.txt
pytest-cov>=2.12.1
pytest-pylint>=0.18.0
yapf
docformatter
isort
# Note: if https://github.com/python/mypy/issues/5485 gets resolved, we can install mypy from head again.
mypy@git+https://github.com/python/mypy.git@9bd651758e8ea2494837814092af70f8d9e6f7a1
